1:1 채팅 리스트

+++ src/main/java/com/daview/mapper/ChatRoomMapper.java
@@ -8,8 +8,8 @@ import java.util.List;
 
 @Mapper
 public interface ChatRoomMapper {
-    List<ChatRoomDTO> getChatRoomListForUser(int memberId);
-    List<ChatRoomDTO> getChatRoomsByMemberId(int memberId);
+    List<ChatRoomDTO> getChatRoomListForUser(Long memberId);
+    List<ChatRoomDTO> getChatRoomsByMemberId(Long memberId);
     
     String findChatRoomId(@Param("senderId") Long senderId, 
                           @Param("receiverId") Long receiverId,
